+++ color.h
@@ -6,12 +6,12 @@
 
 class Color : public Tuple {
  public:
-  Color(float red, float green, float blue) : Tuple(red, green, blue, 0.0) {}
+  Color(double red, double green, double blue) : Tuple(red, green, blue, 0.0) {}
   Color(const Tuple& a) : Tuple(a.x, a.y, a.z, 0.0) {}
 
-  float r() const { return x; }
-  float g() const { return y; }
-  float b() const { return z; }
+  double r() const { return x; }
+  double g() const { return y; }
+  double b() const { return z; }
 
   friend std::ostream& operator<<(std::ostream& os, const Color& rhs) {
     return os << "Color(" << rhs.x << ", " << rhs.y << ", " << rhs.z << ")";
